# docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: ボットちゃんねる API
  version: 0.1.0
  description: |
    ボットちゃんねる の Web API 仕様。
    5ch専ブラ互換APIは別セクション（paths末尾）に記載。

    ## 認証方式
    - 一般ユーザー: edge-token Cookie（書き込みAPI）
    - 管理者: admin_session Cookie（管理API）
    - ボット: X-Bot-API-Key ヘッダ（書き込みAPI）

    ## ステータス
    ドラフト（Phase 1 + Phase 2 MVP）
servers:
  - url: https://{domain}
    description: 本番環境（Vercel / Cloudflare、両環境並行稼働）
    variables:
      domain:
        default: battleboard.vercel.app

# =============================================
# セキュリティ定義
# =============================================
components:
  securitySchemes:
    edgeTokenCookie:
      type: apiKey
      in: cookie
      name: edge-token
      description: 一般ユーザーの書き込み認証トークン

    adminSessionCookie:
      type: apiKey
      in: cookie
      name: admin_session
      description: 管理者セッション Cookie

    botApiKey:
      type: apiKey
      in: header
      name: X-Bot-API-Key
      description: 運営ボット用サービスアカウントAPIキー

    internalBearerAuth:
      type: http
      scheme: bearer
      description: |
        Internal API 専用認証。GitHub Actions cron ジョブのみが呼び出す。
        Authorization: Bearer {BOT_API_KEY} ヘッダで送信する。
        BOT_API_KEY は環境変数で管理される（クライアントサイドに含めることを禁止）。

  # =============================================
  # スキーマ定義
  # =============================================
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        threadId:
          type: string
          format: uuid
        postNumber:
          type: integer
          minimum: 1
          example: 5
        displayName:
          type: string
          example: 名無しさん
        dailyId:
          type: string
          example: Ax8kP2Lm
        body:
          type: string
        isSystemMessage:
          type: boolean
        isDeleted:
          type: boolean
        inlineSystemInfo:
          type: string
          nullable: true
          description: レス内マージ型システム情報（コマンド結果・書き込み報酬等）。方式A表示用。
        botMark:
          type: object
          nullable: true
          description: BOTマーク情報（告発成功後のみ）
          properties:
            hp:
              type: integer
            maxHp:
              type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, threadId, postNumber, displayName, dailyId, body, isSystemMessage, isDeleted, createdAt]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
paths:
...
  /api/admin/users:
    get:
      operationId: adminListUsers
      summary: ユーザー一覧取得（管理者）
      description: |
        ユーザー一覧をページネーション付きで取得する。
      tags: [管理者]
      security:
        - adminSessionCookie: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [created_at, last_post_date]
            default: created_at
      responses:
        "200":
          description: ユーザー一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      description: User エンティティ
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        "401":
          description: 管理者権限なし
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users/{userId}:
    get:
      operationId: adminGetUserDetail
      summary: ユーザー詳細取得（管理者）
      description: |
        指定ユーザーの詳細情報（基本情報・通貨残高・書き込み履歴最新50件）を取得する。
      tags: [管理者]
      security:
        - adminSessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: ユーザー詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  createdAt:
                    type: string
                    format: date-time
                  isBanned:
                    type: boolean
                  isPremium:
                    type: boolean
                  registrationType:
                    type: string
                    nullable: true
                    enum: [email, discord, null]
                  username:
                    type: string
                    nullable: true
                  streakDays:
                    type: integer
                  grassCount:
                    type: integer
                  balance:
                    type: integer
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
        "401":
          description: 管理者権限なし
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: ユーザーが存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users/{userId}/posts:
    get:
      operationId: adminGetUserPosts
      summary: ユーザー書き込み履歴取得（管理者）
      description: |
        指定ユーザーの書き込み履歴をページネーション付きで取得する（created_at DESC）。
      tags: [管理者]
      security:
        - adminSessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: 書き込み履歴
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
                  limit:
                    type: integer
                  offset:
                    type: integer
        "401":
          description: 管理者権限なし
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # -----------------------------------------
  # 管理者: ダッシュボード
  # -----------------------------------------
  /api/admin/dashboard:
    get:
      operationId: adminGetDashboard
      summary: ダッシュボード統計情報取得（管理者）
      description: |
        ダッシュボードのリアルタイムサマリーを取得する。
        本日分はリアルタイム集計（UserRepository, PostRepository, CurrencyRepository から直接集計）。
      tags: [管理者]
      security:
        - adminSessionCookie: []
      parameters:
        - name: today
          in: query
          required: false
          schema:
            type: string
            format: date
          description: 本日日付（YYYY-MM-DD）。省略時はサーバー現在日付
      responses:
        "200":
          description: ダッシュボードサマリー
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUsers:
                    type: integer
                    description: 総ユーザー数（仮+本登録）
                  todayPosts:
                    type: integer
                    description: 本日の書き込み数（非システムメッセージ）
                  activeThreads:
                    type: integer
                    description: アクティブスレッド数（本日書き込みがあったスレッド）
                  currencyInCirculation:
                    type: integer
                    description: 通貨流通量（全ユーザー残高合計）
                required: [totalUsers, todayPosts, activeThreads, currencyInCirculation]
        "401":
          description: 管理者権限なし
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/dashboard/history:
    get:
      operationId: adminGetDashboardHistory
      summary: ダッシュボード日次推移取得（管理者）
      description: |
        ダッシュボードの日次推移を daily_stats テーブルから取得する。
        本日分はリアルタイム集計のため /dashboard を参照する。
      tags: [管理者]
      security:
        - adminSessionCookie: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
            maximum: 90
          description: 取得日数
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: 開始日（YYYY-MM-DD）。省略時は today - days
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: 終了日（YYYY-MM-DD）。省略時は昨日
      responses:
        "200":
          description: 日次推移データ
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        statDate:
                          type: string
                          format: date
                        totalUsers:
                          type: integer
                        newUsers:
                          type: integer
                        activeUsers:
                          type: integer
                        totalPosts:
                          type: integer
                        totalThreads:
                          type: integer
                        activeThreads:
                          type: integer
                        currencyInCirculation:
                          type: integer
                        currencyGranted:
                          type: integer
                        currencyConsumed:
                          type: integer
                        totalAccusations:
                          type: integer
                        totalAttacks:
                          type: integer
                  days:
                    type: integer
        "401":
          description: 管理者権限なし
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # -----------------------------------------
  # Internal API（サーバー間通信専用）
  # Bearer認証: Authorization: Bearer {BOT_API_KEY}
  # 呼び出し元: GitHub Actions cronジョブのみ
  # -----------------------------------------
...